requirements.txt
dataset
requests
pandas
h5py
keras
sklearn
tensorflow
redis==2.10.6
python-dotenv==0.8.2